hubble/parser/seven: split L7 protocol handling to specific source files
There are no functional changes, only code moved around to new files.
